+++ llvm-hs/src/LLVM/Internal/FFI/OrcJITC.cpp
@@ -1,35 +1,211 @@
 #include "LLVM/Internal/FFI/OrcJIT.h"
 #include "LLVM/Internal/FFI/Target.hpp"
+#include "llvm/ExecutionEngine/JITSymbol.h"
 #include "llvm/ExecutionEngine/Orc/CompileOnDemandLayer.h"
 #include "llvm/ExecutionEngine/Orc/CompileUtils.h"
 #include "llvm/ExecutionEngine/Orc/IRCompileLayer.h"
 #include "llvm/ExecutionEngine/Orc/IndirectionUtils.h"
-#include "llvm/ExecutionEngine/JITSymbol.h"
 #include "llvm/ExecutionEngine/Orc/LambdaResolver.h"
 #include "llvm/ExecutionEngine/Orc/ObjectLinkingLayer.h"
 #include "llvm/IR/Mangler.h"
 
 #include <type_traits>
+#include <unordered_map>
 
 using namespace llvm;
 using namespace orc;
 
-typedef llvm::orc::ObjectLinkingLayer<> *LLVMObjectLinkingLayerRef;
-typedef llvm::orc::IRCompileLayer<llvm::orc::ObjectLinkingLayer<>>
-    LLVMIRCompileLayer;
-typedef LLVMIRCompileLayer *LLVMIRCompileLayerRef;
-typedef llvm::orc::CompileOnDemandLayer<LLVMIRCompileLayer>
-    LLVMCompileOnDemandLayer;
-typedef LLVMCompileOnDemandLayer *LLVMCompileOnDemandLayerRef;
-typedef llvm::orc::JITCompileCallbackManager *LLVMJITCompileCallbackManagerRef;
-typedef llvm::JITSymbol *LLVMJITSymbolRef;
-typedef LLVMIRCompileLayer::ModuleSetHandleT *LLVMModuleSetHandleRef;
-typedef LLVMCompileOnDemandLayer::ModuleSetHandleT *LLVMCODModuleSetHandleRef;
+typedef unsigned LLVMModuleSetHandle;
+typedef unsigned LLVMObjSetHandle;
 typedef llvm::orc::LambdaResolver<
     std::function<JITSymbol(const std::string &name)>,
     std::function<JITSymbol(const std::string &name)>>
     LLVMLambdaResolver;
 typedef LLVMLambdaResolver *LLVMLambdaResolverRef;
+
+// We want to allow users to choose themselves which layers they want to use.
+// However, the LLVM API requires that this is selected statically via template
+// arguments. We convert this static polymorphism to runtime polymorphism by
+// creating an ObjectLayer and a CompileLayer class which use virtual dispatch
+// to select the concrete layer.
+
+template <typename T> class HandleSet {
+  public:
+    unsigned insert(T t) {
+        unsigned handle = handles.insert({nextFree, t}).first->first;
+        ++nextFree;
+        return handle;
+    }
+    T &lookup(unsigned i) { return handles.at(i); }
+
+    void remove(unsigned i) { handles.erase(i); }
+
+  private:
+    std::unordered_map<unsigned, T> handles;
+    unsigned nextFree = 0;
+};
+
+class ObjectLayer {
+  public:
+    virtual ~ObjectLayer(){};
+    typedef unsigned ObjSetHandleT;
+    virtual ObjSetHandleT
+    addObjectSet(std::vector<std::unique_ptr<object::ObjectFile>> objects,
+                 SectionMemoryManager *memMgr,
+                 LLVMLambdaResolverRef resolver) = 0;
+    virtual ObjSetHandleT addObjectSet(
+        std::vector<std::unique_ptr<object::OwningBinary<object::ObjectFile>>>
+            objects,
+        SectionMemoryManager *memMgr, LLVMLambdaResolverRef resolver) = 0;
+    virtual ObjSetHandleT addObjectSet(
+        std::vector<std::unique_ptr<object::OwningBinary<object::ObjectFile>>>
+            objects,
+        std::unique_ptr<SectionMemoryManager> memMgr,
+        LLVMLambdaResolverRef resolver) = 0;
+    virtual ObjSetHandleT addObjectSet(
+        std::vector<std::unique_ptr<object::OwningBinary<object::ObjectFile>>>
+            objects,
+        std::unique_ptr<SectionMemoryManager> memMgr,
+        std::unique_ptr<LLVMLambdaResolver> resolver) = 0;
+    virtual ObjSetHandleT addObjectSet(
+        std::vector<std::unique_ptr<object::OwningBinary<object::ObjectFile>>>
+            objects,
+        SectionMemoryManager *memMgr,
+        std::unique_ptr<JITSymbolResolver> resolver) = 0;
+    virtual void removeObjectSet(ObjSetHandleT H) = 0;
+    virtual JITSymbol findSymbol(StringRef name, bool exportedSymbolsOnly) = 0;
+    virtual JITSymbol findSymbolIn(ObjSetHandleT h, StringRef name,
+                                   bool exportedSymbolsOnly) = 0;
+    virtual void emitAndFinalize(ObjSetHandleT h) = 0;
+};
+
+template <typename T> class ObjectLayerT : public ObjectLayer {
+  public:
+    ObjectLayerT(T data_) : data(std::move(data_)) {}
+    ObjSetHandleT
+    addObjectSet(std::vector<std::unique_ptr<object::ObjectFile>> objects,
+                 SectionMemoryManager *memMgr,
+                 LLVMLambdaResolverRef resolver) override {
+        auto handle = data.addObjectSet(std::move(objects), std::move(memMgr),
+                                        std::move(resolver));
+        return handles.insert(handle);
+    }
+    ObjSetHandleT addObjectSet(
+        std::vector<std::unique_ptr<object::OwningBinary<object::ObjectFile>>>
+            objects,
+        SectionMemoryManager *memMgr, LLVMLambdaResolverRef resolver) override {
+        auto handle = data.addObjectSet(std::move(objects), std::move(memMgr),
+                                        std::move(resolver));
+        return handles.insert(handle);
+    }
+    ObjSetHandleT addObjectSet(
+        std::vector<std::unique_ptr<object::OwningBinary<object::ObjectFile>>>
+            objects,
+        std::unique_ptr<SectionMemoryManager> memMgr,
+        LLVMLambdaResolverRef resolver) override {
+        auto handle = data.addObjectSet(std::move(objects), std::move(memMgr),
+                                        std::move(resolver));
+        return handles.insert(handle);
+    }
+    ObjSetHandleT addObjectSet(
+        std::vector<std::unique_ptr<object::OwningBinary<object::ObjectFile>>>
+            objects,
+        std::unique_ptr<SectionMemoryManager> memMgr,
+        std::unique_ptr<LLVMLambdaResolver> resolver) override {
+        auto handle = data.addObjectSet(std::move(objects), std::move(memMgr),
+                                        std::move(resolver));
+        return handles.insert(handle);
+    }
+    ObjSetHandleT addObjectSet(
+        std::vector<std::unique_ptr<object::OwningBinary<object::ObjectFile>>>
+            objects,
+        SectionMemoryManager *memMgr,
+        std::unique_ptr<JITSymbolResolver> resolver) override {
+        auto handle = data.addObjectSet(std::move(objects), std::move(memMgr),
+                                        std::move(resolver));
+        return handles.insert(handle);
+    }
+    void removeObjectSet(ObjSetHandleT h) override {
+        data.removeObjectSet(handles.lookup(h));
+        handles.remove(h);
+    }
+    JITSymbol findSymbol(StringRef name, bool exportedSymbolsOnly) override {
+        return data.findSymbol(name, exportedSymbolsOnly);
+    }
+    JITSymbol findSymbolIn(ObjSetHandleT h, StringRef name,
+                           bool exportedSymbolsOnly) override {
+        return data.findSymbolIn(handles.lookup(h), name, exportedSymbolsOnly);
+    }
+    void emitAndFinalize(ObjSetHandleT h) override {
+        data.emitAndFinalize(handles.lookup(h));
+    }
+
+  private:
+    T data;
+    HandleSet<typename T::ObjSetHandleT> handles;
+};
+
+class CompileLayer {
+  public:
+    typedef LLVMModuleSetHandle ModuleSetHandleT;
+    virtual ~CompileLayer(){};
+    virtual JITSymbol findSymbol(StringRef Name, bool ExportedSymbolsOnly) = 0;
+    virtual JITSymbol findSymbolIn(ModuleSetHandleT H, StringRef Name,
+                                   bool ExportedSymbolsOnly) = 0;
+    virtual ModuleSetHandleT
+    addModuleSet(std::vector<Module *> Modules,
+                 std::unique_ptr<SectionMemoryManager> MemMgr,
+                 std::unique_ptr<LLVMLambdaResolver> Resolver) = 0;
+    virtual ModuleSetHandleT
+    addModuleSet(std::vector<std::unique_ptr<Module>> Modules,
+                 SectionMemoryManager *MemMgr,
+                 std::unique_ptr<JITSymbolResolver> Resolver) = 0;
+    virtual void removeModuleSet(ModuleSetHandleT H) = 0;
+};
+
+template <typename T> class CompileLayerT : public CompileLayer {
+  public:
+    CompileLayerT(T data_) : data(std::move(data_)) {}
+    JITSymbol findSymbol(StringRef Name, bool ExportedSymbolsOnly) override {
+        return data.findSymbol(Name, ExportedSymbolsOnly);
+    }
+    JITSymbol findSymbolIn(ModuleSetHandleT H, StringRef Name,
+                           bool ExportedSymbolsOnly) override {
+        return data.findSymbolIn(handles.lookup(H), Name, ExportedSymbolsOnly);
+    }
+    ModuleSetHandleT
+    addModuleSet(std::vector<Module *> Modules,
+                 std::unique_ptr<SectionMemoryManager> MemMgr,
+                 std::unique_ptr<LLVMLambdaResolver> Resolver) override {
+        auto handle = data.addModuleSet(std::move(Modules), std::move(MemMgr),
+                                        std::move(Resolver));
+        return handles.insert(handle);
+    }
+    ModuleSetHandleT
+    addModuleSet(std::vector<std::unique_ptr<Module>> Modules,
+                 SectionMemoryManager *MemMgr,
+                 std::unique_ptr<JITSymbolResolver> Resolver) override {
+        auto handle = data.addModuleSet(std::move(Modules), std::move(MemMgr),
+                                        std::move(Resolver));
+        return handles.insert(handle);
+    }
+    void removeModuleSet(ModuleSetHandleT H) override {
+        data.removeModuleSet(handles.lookup(H));
+        handles.remove(H);
+    }
+
+  private:
+    T data;
+    HandleSet<typename T::ModuleSetHandleT> handles;
+};
+
+typedef llvm::orc::CompileOnDemandLayer<CompileLayer> LLVMCompileOnDemandLayer;
+typedef LLVMCompileOnDemandLayer *LLVMCompileOnDemandLayerRef;
+
+typedef llvm::orc::JITCompileCallbackManager *LLVMJITCompileCallbackManagerRef;
+
+typedef llvm::JITSymbol *LLVMJITSymbolRef;
+
 typedef llvm::orc::IndirectStubsManager *LLVMIndirectStubsManagerRef;
 typedef std::function<std::unique_ptr<llvm::orc::IndirectStubsManager>()>
     *LLVMIndirectStubsManagerBuilderRef;
@@ -58,24 +234,74 @@ static std::vector<Module *> getModules(LLVMModuleRef *modules,
 }
 
 extern "C" {
-LLVMIRCompileLayerRef
-LLVM_Hs_createIRCompileLayer(LLVMObjectLinkingLayerRef objectLayer,
-                                  LLVMTargetMachineRef tm) {
+
+/* Constructor functions for the different compile layers */
+
+CompileLayer *LLVM_Hs_createIRCompileLayer(ObjectLayer *objectLayer,
+                                           LLVMTargetMachineRef tm) {
     TargetMachine *tmm = unwrap(tm);
-    return new IRCompileLayer<ObjectLinkingLayer<>>(*objectLayer,
-                                                    SimpleCompiler(*tmm));
+    return new CompileLayerT<IRCompileLayer<ObjectLayer>>(
+        IRCompileLayer<ObjectLayer>(*objectLayer, SimpleCompiler(*tmm)));
 }
 
-void LLVM_Hs_disposeIRCompileLayer(LLVMIRCompileLayerRef compileLayer) {
+CompileLayer *LLVM_Hs_createCompileOnDemandLayer(
+    CompileLayer *compileLayer,
+    void (*partitioningFtor)(llvm::Function *, std::set<llvm::Function *> *set),
+    LLVMJITCompileCallbackManagerRef callbackManager,
+    LLVMIndirectStubsManagerBuilderRef stubsManager,
+    LLVMBool cloneStubsIntoPartitions) {
+    return new CompileLayerT<LLVMCompileOnDemandLayer>(LLVMCompileOnDemandLayer(
+        *compileLayer,
+        [partitioningFtor](llvm::Function &f) -> std::set<llvm::Function *> {
+            std::set<llvm::Function *> result;
+            partitioningFtor(&f, &result);
+            return result;
+        },
+        *callbackManager, *stubsManager, cloneStubsIntoPartitions));
+}
+
+/* Functions that work on all compile layers */
+
+void LLVM_Hs_CompileLayer_dispose(CompileLayer *compileLayer) {
     delete compileLayer;
 }
 
-LLVMJITSymbolRef
-LLVM_Hs_IRCompileLayer_findSymbol(LLVMIRCompileLayerRef compileLayer,
-                                       const char *name,
-                                       LLVMBool exportedSymbolsOnly) {
+LLVMJITSymbolRef LLVM_Hs_CompileLayer_findSymbol(CompileLayer *compileLayer,
+                                                 const char *name,
+                                                 LLVMBool exportedSymbolsOnly) {
     JITSymbol symbol = compileLayer->findSymbol(name, exportedSymbolsOnly);
     return new JITSymbol(symbol);
+}
+
+LLVMModuleSetHandle
+LLVM_Hs_CompileLayer_addModuleSet(CompileLayer *compileLayer,
+                                  LLVMTargetDataRef dataLayout,
+                                  LLVMModuleRef *modules, unsigned moduleCount,
+                                  LLVMLambdaResolverRef resolver) {
+    std::vector<Module *> moduleVec =
+        getModules(modules, moduleCount, dataLayout);
+    std::unique_ptr<LLVMLambdaResolver> uniqueResolver(
+        new LLVMLambdaResolver(*resolver));
+    return compileLayer->addModuleSet(moduleVec,
+                                      make_unique<SectionMemoryManager>(),
+                                      std::move(uniqueResolver));
+}
+
+void LLVM_Hs_CompileLayer_removeModuleSet(CompileLayer *compileLayer,
+                                          LLVMModuleSetHandle moduleSetHandle) {
+    compileLayer->removeModuleSet(moduleSetHandle);
+}
+
+/* Constructor functions for the different object layers */
+
+ObjectLayer *LLVM_Hs_createObjectLinkingLayer() {
+    return new ObjectLayerT<ObjectLinkingLayer<>>(ObjectLinkingLayer<>());
+}
+
+/* Fuctions that work on all object layers */
+
+void LLVM_Hs_ObjectLayer_dispose(ObjectLayer *objectLayer) {
+    delete objectLayer;
 }
 
 void LLVM_Hs_disposeJITSymbol(LLVMJITSymbolRef symbol) { delete symbol; }
@@ -83,16 +309,14 @@ void LLVM_Hs_disposeJITSymbol(LLVMJITSymbolRef symbol) { delete symbol; }
 LLVMLambdaResolverRef LLVM_Hs_createLambdaResolver(
     void (*dylibResolver)(const char *, LLVMJITSymbolRef),
     void (*externalResolver)(const char *, LLVMJITSymbolRef)) {
-    std::function<JITSymbol(const std::string &name)>
-        dylibResolverFun = [dylibResolver](
-            const std::string &name) -> JITSymbol {
+    std::function<JITSymbol(const std::string &name)> dylibResolverFun =
+        [dylibResolver](const std::string &name) -> JITSymbol {
         JITSymbol symbol(nullptr);
         dylibResolver(name.c_str(), &symbol);
         return symbol;
     };
-    std::function<JITSymbol(const std::string &name)>
-        externalResolverFun = [externalResolver](
-            const std::string &name) -> JITSymbol {
+    std::function<JITSymbol(const std::string &name)> externalResolverFun =
+        [externalResolver](const std::string &name) -> JITSymbol {
         JITSymbol symbol(nullptr);
         externalResolver(name.c_str(), &symbol);
         return symbol;
@@ -100,33 +324,6 @@ LLVMLambdaResolverRef LLVM_Hs_createLambdaResolver(
     auto lambdaResolver =
         createLambdaResolver(dylibResolverFun, externalResolverFun);
     return lambdaResolver.release();
-}
-
-LLVMModuleSetHandleRef LLVM_Hs_IRCompileLayer_addModuleSet(
-    LLVMIRCompileLayerRef compileLayer, LLVMTargetDataRef dataLayout,
-    LLVMModuleRef *modules, unsigned moduleCount,
-    LLVMLambdaResolverRef resolver) {
-    std::vector<Module *> moduleVec =
-        getModules(modules, moduleCount, dataLayout);
-    return new IRCompileLayer<ObjectLinkingLayer<>>::ModuleSetHandleT(
-        compileLayer->addModuleSet(
-            moduleVec, make_unique<SectionMemoryManager>(), resolver));
-}
-
-void LLVM_Hs_IRCompileLayer_removeModuleSet(
-    LLVMIRCompileLayerRef compileLayer,
-    LLVMModuleSetHandleRef moduleSetHandle) {
-    compileLayer->removeModuleSet(*moduleSetHandle);
-    delete moduleSetHandle;
-}
-
-LLVMObjectLinkingLayerRef LLVM_Hs_createObjectLinkingLayer() {
-    return new ObjectLinkingLayer<>();
-}
-
-void LLVM_Hs_disposeObjectLinkingLayer(
-    LLVMObjectLinkingLayerRef objectLayer) {
-    delete objectLayer;
 }
 
 static JITSymbolFlags unwrap(LLVMJITSymbolFlags f) {
@@ -149,8 +346,7 @@ static LLVMJITSymbolFlags wrap(JITSymbolFlags f) {
     return LLVMJITSymbolFlags(r);
 }
 
-JITTargetAddress
-LLVM_Hs_JITSymbol_getAddress(LLVMJITSymbolRef symbol) {
+JITTargetAddress LLVM_Hs_JITSymbol_getAddress(LLVMJITSymbolRef symbol) {
     return symbol->getAddress();
 }
 
@@ -158,14 +354,13 @@ LLVMJITSymbolFlags LLVM_Hs_JITSymbol_getFlags(LLVMJITSymbolRef symbol) {
     return wrap(symbol->getFlags());
 }
 
-void LLVM_Hs_setJITSymbol(LLVMJITSymbolRef symbol,
-                               JITTargetAddress addr,
-                               LLVMJITSymbolFlags flags) {
+void LLVM_Hs_setJITSymbol(LLVMJITSymbolRef symbol, JITTargetAddress addr,
+                          LLVMJITSymbolFlags flags) {
     *symbol = JITSymbol(addr, unwrap(flags));
 }
 
 void LLVM_Hs_getMangledSymbol(char **mangledSymbol, const char *symbol,
-                                   LLVMTargetDataRef dataLayout) {
+                              LLVMTargetDataRef dataLayout) {
     std::string mangled = mangle(symbol, dataLayout);
     *mangledSymbol = new char[mangled.size() + 1];
     strcpy(*mangledSymbol, mangled.c_str());
@@ -175,8 +370,9 @@ void LLVM_Hs_disposeMangledSymbol(char *mangledSymbol) {
     delete[] mangledSymbol;
 }
 
-LLVMJITCompileCallbackManagerRef LLVM_Hs_createLocalCompileCallbackManager(
-    const char *triple, JITTargetAddress errorHandler) {
+LLVMJITCompileCallbackManagerRef
+LLVM_Hs_createLocalCompileCallbackManager(const char *triple,
+                                          JITTargetAddress errorHandler) {
     return llvm::orc::createLocalCompileCallbackManager(Triple(triple),
                                                         errorHandler)
         .release();
@@ -198,59 +394,7 @@ void LLVM_Hs_disposeIndirectStubsManagerBuilder(
     delete stubsManager;
 }
 
-void LLVM_Hs_insertFun(std::set<llvm::Function *> *set,
-                            llvm::Function *f) {
+void LLVM_Hs_insertFun(std::set<llvm::Function *> *set, llvm::Function *f) {
     set->insert(f);
-}
-
-LLVMCompileOnDemandLayerRef LLVM_Hs_createCompileOnDemandLayer(
-    LLVMIRCompileLayerRef compileLayer,
-    void (*partitioningFtor)(llvm::Function *, std::set<llvm::Function *> *set),
-    LLVMJITCompileCallbackManagerRef callbackManager,
-    LLVMIndirectStubsManagerBuilderRef stubsManager,
-    LLVMBool cloneStubsIntoPartitions) {
-    return new LLVMCompileOnDemandLayer(
-        *compileLayer,
-        [partitioningFtor](llvm::Function &f) -> std::set<llvm::Function *> {
-            std::set<llvm::Function *> result;
-            partitioningFtor(&f, &result);
-            return result;
-        },
-        *callbackManager, *stubsManager, cloneStubsIntoPartitions);
-}
-
-void LLVM_Hs_disposeCompileOnDemandLayer(
-    LLVMCompileOnDemandLayerRef codLayer) {
-    delete codLayer;
-}
-
-LLVMCODModuleSetHandleRef LLVM_Hs_CompileOnDemandLayer_addModuleSet(
-    LLVMCompileOnDemandLayerRef compileLayer, LLVMTargetDataRef dataLayout,
-    LLVMModuleRef *modules, unsigned moduleCount,
-    LLVMLambdaResolverRef resolver) {
-    std::vector<Module *> moduleVec =
-        getModules(modules, moduleCount, dataLayout);
-    // We need to copy the resolver to make the use of unique_ptr (required by
-    // the LLVM API) safe
-    std::unique_ptr<LLVMLambdaResolver> uniqueResolver(
-        new LLVMLambdaResolver(*resolver));
-    return new LLVMCompileOnDemandLayer::ModuleSetHandleT(
-        compileLayer->addModuleSet(moduleVec,
-                                   make_unique<SectionMemoryManager>(),
-                                   std::move(uniqueResolver)));
-}
-
-void LLVM_Hs_CompileOnDemandLayer_removeModuleSet(
-    LLVMCompileOnDemandLayerRef compileLayer,
-    LLVMCODModuleSetHandleRef moduleSetHandle) {
-    compileLayer->removeModuleSet(*moduleSetHandle);
-    delete moduleSetHandle;
-}
-
-LLVMJITSymbolRef LLVM_Hs_CompileOnDemandLayer_findSymbol(
-    LLVMCompileOnDemandLayerRef compileLayer, const char *name,
-    LLVMBool exportedSymbolsOnly) {
-    JITSymbol symbol = compileLayer->findSymbol(name, exportedSymbolsOnly);
-    return new JITSymbol(symbol);
 }
 }