+++ llvm-hs/src/LLVM/OrcJIT/IRCompileLayer.hs
@@ -1,10 +1,12 @@
 module LLVM.OrcJIT.IRCompileLayer (
     IRCompileLayer,
-    ModuleSet,
+    ModuleSetHandle,
     findSymbol,
     mangleSymbol,
     withIRCompileLayer,
     withModuleSet
   ) where
 
-import           LLVM.Internal.OrcJIT.IRCompileLayer
+import LLVM.Internal.OrcJIT.CompileLayer
+import LLVM.Internal.OrcJIT.IRCompileLayer
+import LLVM.Internal.FFI.OrcJIT.CompileLayer (ModuleSetHandle)